pyproject.toml
[tool.poetry]
name = "the-factory-bot-task"
version = "0.1.0"
description = ""
authors = ["K-hage"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
aiogram = "^3.1.1"
envparse = "^0.2.0"
psycopg2-binary = "^2.9.8"
drf-spectacular = "^0.26.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
